+++ src/lib/databases/Get-CosmosDbDatabaseResourcePath.ps1
@@ -6,7 +6,7 @@ function Get-CosmosDbDatabaseResourcePath
     param
     (
         [Parameter(Mandatory = $true)]
-        [ValidateNotNullOrEmpty()]
+        [ValidateScript({ Assert-CosmosDbDatabaseIdValid -Id $_ })]
         [System.String]
         $Id
     )