+++ src/lib/utils/New-CosmosDbContext.ps1
@@ -9,12 +9,12 @@ function New-CosmosDbContext
         [Parameter(Mandatory = $true, ParameterSetName = 'Account')]
         [Parameter(Mandatory = $true, ParameterSetName = 'Token')]
         [Parameter(Mandatory = $true, ParameterSetName = 'AzureAccount')]
-        [ValidateNotNullOrEmpty()]
+        [ValidateScript({ Assert-CosmosDbAccountNameValid -Name $_ })]
         [System.String]
         $Account,
 
         [Parameter()]
-        [ValidateNotNullOrEmpty()]
+        [ValidateScript({ Assert-CosmosDbDatabaseIdValid -Id $_ })]
         [System.String]
         $Database,